+++ test/test_develop.py
@@ -22,8 +22,6 @@ class TempDevelopTest(_SyncTestBase):
     def setUp(self):
         # Call self._prepare_initial_synced_fixture():
         super(TempDevelopTest, self).setUp()
-        if self.__class__ is TempDevelopTest:
-            raise SkipTest("Only FTP")
         self.local_url = get_local_test_url()
         self.remote_url = get_remote_test_url()
 
@@ -47,36 +45,9 @@ class TempDevelopTest(_SyncTestBase):
         ftp_downloader.run()
         ftp_uploader.run()
 
-#         stats = self.do_run_suite(DownloadSynchronizer, opts)
-#         self._dump_de_facto_results(stats)
-#
-#         # We expect 7 conflicts, and leave them unresolved (i.e. skip them all)
-#
-#         # self.assertEqual(stats["files_written"], 3)
-#         # self.assertEqual(stats["files_written"], 3)
-#         # self.assertEqual(stats["download_files_written"], 0)
-#         # self.assertEqual(stats["upload_files_written"], 3)
-#         # self.assertEqual(stats["files_deleted"], 1)
-#         # self.assertEqual(stats["dirs_deleted"], 2)
-#         # self.assertEqual(stats["conflict_files"], 7)
-#         # self.assertEqual(stats["conflict_files_skipped"], 7)
-#
-#         # We expect that remote remains unmodified
-#         self.assert_test_folder_equal(get_test_folder("remote"),
-#                                       _SyncTestBase.remote_fixture_modified)
-#
-#         # We expect that local only contains files that match '*1.txt'
-#         expect_local = {
-#             'file1.txt': {'content': 'local1', 'date': '2014-01-01 12:00:00'},
-#             'folder1/file1_1.txt': {'content': 'local1_1', 'date': '2014-01-01 12:00:00'},
-#             'folder2/file2_1.txt': {'content': 'local2_1', 'date': '2014-01-01 12:00:00'},
-#             'folder3/file3_1.txt': {'content': 'local3_1', 'date': '2014-01-01 12:00:00'},
-#             'folder4/file4_1.txt': {'content': 'remote 13:00', 'date': '2014-01-01 13:00:00'},
-#             'folder5/file5_1.txt': {'content': 'remote 13:00', 'date': '2014-01-01 13:00:00'},
-#             }
-#         self.assert_test_folder_equal(get_test_folder("local"), expect_local)
-
     def test_issue_21(self):
+        if not self.use_ftp_target:
+            raise SkipTest("Only FTP targets.")
         write_test_file("local/large.txt", size=10*1000)
         write_test_file("remote/large.txt", size=10*1000)
 
@@ -89,6 +60,10 @@ class TempDevelopTest(_SyncTestBase):
 #         print(out)
         assert "*cmd* 'PORT" in out or "*cmd* 'EPRT" in out
         assert not ("*cmd* 'PASV" in out or "*cmd* 'EPSV" in out)
+
+    def test_issue_22(self):
+        # write()
+        pass
 
 
 # ===============================================================================